pyproject.toml
[project]
name = "sturdr"
version = "0.1.0"
description = "Daniel Sturdivant's Software Defined Radio"
authors = [
    { name = "sturdivant20", email = "sturdivant20@gmail.com" }
]
dependencies = [
    "numpy>=2.1.1",
    "matplotlib>=3.9.2",
    "pyqt6>=6.7.1",
    "ipykernel>=6.29.5",
    "sphinx>=8.0.2",
    "numpydoc>=1.8.0",
    "pydata-sphinx-theme>=0.15.4",
    "pyyaml>=6.0.2",
    "numba>=0.60.0",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["sturdr"]
